EES-3799 - only saving items that have not yet been saved.
During the saving of Subject Meta, we explicitly only save the items that have no ids (and therefore are not yet persisted).  During the saving of Filter Items, we give them the tracked FilterGroup rather than the untracked one if it is already persisted, so that the saving of the FilterItem does not then attempt to cascade insert the untracked FilterGroup (as it appears to the DbContext to be unpersisted).